move sulten views and serializers
